feat: project is now callable by 'speaker_verification' module
